Refactor: 서비스 계층 DIP
