+++ src/main/java/com/springcooler/sgma/studygroupmember/command/domain/service/DomainStudyGroupMemberService.java
@@ -1,13 +1,7 @@
 package com.springcooler.sgma.studygroupmember.command.domain.service;
 
-import com.springcooler.sgma.studygroupmember.command.domain.aggregate.StudyGroupMemberStatus;
-import org.springframework.stereotype.Service;
+public interface DomainStudyGroupMemberService {
 
-@Service
-public class DomainStudyGroupMemberService {
-
-    public boolean isActive(String memberStatus) {
-        return memberStatus.equals(StudyGroupMemberStatus.ACTIVE.name());
-    }
+    boolean isActive(String memberStatus);
 
 }
