+++ src/main/java/com/springcooler/sgma/studygroupmember/command/domain/service/DomainStudyGroupMemberServiceImpl.java
@@ -1,0 +1,14 @@
+package com.springcooler.sgma.studygroupmember.command.domain.service;
+
+import com.springcooler.sgma.studygroupmember.command.domain.aggregate.StudyGroupMemberStatus;
+import org.springframework.stereotype.Service;
+
+@Service
+public class DomainStudyGroupMemberServiceImpl implements DomainStudyGroupMemberService {
+
+    @Override
+    public boolean isActive(String memberStatus) {
+        return memberStatus.equals(StudyGroupMemberStatus.ACTIVE.name());
+    }
+
+}
